src/requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.145
confluent-kafka==2.5.0
faker==26.0.0
fastavro==1.9.5
flask-login==0.6.3
flask==3.0.3
langchain-aws==0.1.11
langchain-community==0.2.7
langchain-core==0.2.22
langchain-groq==0.1.6
langchain-openai==0.1.17
langchain==0.2.9
python-dotenv==1.0.1
qdrant-client==1.10.1
requests==2.32.3
sentence-transformers==3.0.1
tiktoken==0.7.0
psycopg2-binary==2.9.9
passlib==1.7.4
fastapi==0.114.2
uvicorn==0.30.6